requirements.txt
tensorflow
pandas
numpy
opencv-python
Pillow
matplotlib